+++ source/frontend/StarChatBubbleManager.hpp
@@ -55,9 +55,9 @@ private:
   // edge of screen
   uint8_t calcDistanceFadeAlpha(Vec2F bubbleScreenPosition, StoredFunctionPtr fadeFunction) const;
 
-  RectF bubbleImageRect(Vec2F screenPos, BubbleImage const& bubbleImage, int pixelRatio);
-  void drawBubbleImage(Vec2F screenPos, BubbleImage const& bubbleImage, int pixelRatio, int alpha);
-  void drawBubbleText(Vec2F screenPos, BubbleText const& bubbleText, int pixelRatio, int alpha, bool isPortrait);
+  RectF bubbleImageRect(Vec2F screenPos, BubbleImage const& bubbleImage, float pixelRatio);
+  void drawBubbleImage(Vec2F screenPos, BubbleImage const& bubbleImage, float pixelRatio, int alpha);
+  void drawBubbleText(Vec2F screenPos, BubbleText const& bubbleText, float pixelRatio, int alpha, bool isPortrait);
 
   GuiContext* m_guiContext;
 
@@ -70,7 +70,7 @@ private:
   Vec2F m_textPadding;
 
   BubbleSeparator<Bubble> m_bubbles;
-  int m_zoom;
+  float m_zoom;
   Vec2F m_bubbleOffset;
   float m_maxAge;
   float m_portraitMaxAge;
